requirements.txt
azure-functions
pymongo
requests
pandas
